+++ include/pangolin/handler/handler_image.h
@@ -59,7 +59,7 @@ public:
 
     void glRenderOverlay();
 
-    void ScreenToImage(Viewport& v, int xpix, int ypix, float& ximg, float& yimg);
+    void ScreenToImage(Viewport& v, float xpix, float ypix, float& ximg, float& yimg);
 
     void ImageToScreen(Viewport& v, float ximg, float yimg, float& xpix, float& ypix);
