+++ include/pangolin/video/drivers/thread.h
@@ -146,12 +146,12 @@ public:
         emptyBuffers.push_back(bp);
     }
 
-    const unsigned int AvailableFrames() const {
+    const size_t AvailableFrames() const {
         boostd::lock_guard<boostd::mutex> vlock(vMtx);
         return validBuffers.size();
     }
 
-    const unsigned int EmptyBuffers() const {
+    const size_t EmptyBuffers() const {
         boostd::lock_guard<boostd::mutex> elock(eMtx);
         return emptyBuffers.size();
     }
